commit22

+++ src/app/components/country/country.component.ts
@@ -43,7 +43,10 @@ export class CountryComponent implements OnInit {
 }
 
   public deleteCountry(id){
+    let a = confirm("Are you Sure you want to delete this Brand ?")
+    if (a == true){
     this._dataService.deleteCountry({"id": id})
+    }
   }
 
 }
